Fix Shadow Masks Textures Having A 1 Frame Creation Delay (#5335)
* GetOrCreateShadowMask() API for shadow masks, simpler to get a per-view texture for the mask and is resolved on the same frame, fixes 1 frame delay on screenspace shadows and VR rendering of them

* Start cleaning up this ShadowMask API

* Move shadow mask stuff to command list, the way we do SSS still needs some stuff we cant nicely expose so use AddAction

* Remove rendering callback from DirectionalLight, add commandlists to scenelight

* Stash a system for shadow masks so I'll revisit later when to make this public

* Do this much simpler, don't expose API for shadow masks yet, intentionally make screenspace shadows part of shadowmapper

* Make this private again and format

* A8 instead of R16

* Only managed cameras reach OnRenderStage, so only they ever clear/dispatch into the mask.

* Use light pointer instead of gethashcode for getting shadow mask

* Init shadowmasktextureindex
